Fixes for Negotiate Auth Protocol
